move playback options
