+++ src/renderer/components/animator/AnimationToolbar/AnimationToolbar.tsx
@@ -1,9 +1,7 @@
 import { useContext, useState } from "react";
 import { useSelector } from "react-redux";
 import { PageRoute } from "../../../../common/PageRoute";
-import PlaybackContext, {
-  PlaybackFrameName,
-} from "../../../context/PlaybackContext/PlaybackContext";
+import PlaybackContext from "../../../context/PlaybackContext/PlaybackContext";
 import { RootState } from "../../../redux/store";
 import { getTrackLength } from "../../../services/project/projectCalculator";
 import IconName from "../../common/Icon/IconName";
@@ -11,31 +9,16 @@ import IconButton from "../../common/IconButton/IconButton";
 import InputRange from "../../common/Input/InputRange/InputRange";
 import Toolbar from "../../common/Toolbar/Toolbar";
 import ToolbarItem, { ToolbarItemAlign } from "../../common/ToolbarItem/ToolbarItem";
-import PlaybackSpeedSelect from "../PlaybackSpeedSelect/PlaybackSpeedSelect";
 import "./AnimationToolbar.css";
 
 export const AnimationToolbar = (): JSX.Element => {
-  const shortPlayLength = useSelector(
-    (state: RootState) => state.app.userPreferences.shortPlayLength
-  );
-  const shortPlayFrameText = shortPlayLength === 1 ? "frame" : "frames";
-
-  const {
-    startOrPausePlayback,
-    stopPlayback,
-    displayFrame,
-    shortPlay,
-    liveViewVisible,
-    timelineIndex,
-    playing,
-  } = useContext(PlaybackContext);
+  const { liveViewVisible, timelineIndex } = useContext(PlaybackContext);
   const frameTrack = useSelector((state: RootState) => state.project.take?.frameTrack);
   if (frameTrack === undefined) {
     throw "No frame track found in AnimationToolbar";
   }
 
   const [onionSkinAmount, setOnionSkinAmount] = useState(0);
-  const [loopPlayback, setLoopPlayback] = useState(false);
 
   return (
     <Toolbar className="animation-toolbar">
@@ -57,50 +40,6 @@ export const AnimationToolbar = (): JSX.Element => {
           max={100}
           step={2}
           value={onionSkinAmount}
-        />
-      </ToolbarItem>
-
-      <ToolbarItem align={ToolbarItemAlign.CENTER}>
-        <IconButton
-          title="First Frame"
-          icon={IconName.PLAY_FIRST}
-          onClick={() => displayFrame(PlaybackFrameName.FIRST)}
-        />
-        <IconButton
-          title="Previous Frame"
-          icon={IconName.PLAY_PREVIOUS}
-          onClick={() => displayFrame(PlaybackFrameName.PREVIOUS)}
-        />
-        <IconButton
-          title={playing ? "Pause Playback" : "Playback Frames"}
-          icon={playing ? IconName.PLAY_PAUSE : IconName.PLAY}
-          onClick={() => startOrPausePlayback()}
-        />
-        <IconButton title="Stop Playback" icon={IconName.PLAY_STOP} onClick={stopPlayback} />
-        <IconButton
-          title="Next Frame"
-          icon={IconName.PLAY_NEXT}
-          onClick={() => displayFrame(PlaybackFrameName.NEXT)}
-        />
-        <IconButton
-          title="Last Frame"
-          icon={IconName.PLAY_LAST}
-          onClick={() => displayFrame(PlaybackFrameName.LAST)}
-        />
-      </ToolbarItem>
-
-      <ToolbarItem stretch align={ToolbarItemAlign.RIGHT}>
-        <PlaybackSpeedSelect />
-        <IconButton
-          title={`Short Play (${shortPlayLength} ${shortPlayFrameText})`}
-          icon={IconName.PLAY_SHORT}
-          onClick={shortPlay}
-        />
-        <IconButton
-          title={`${loopPlayback ? "Disable" : "Enable"} Loop Playback`}
-          icon={IconName.PLAY_LOOP}
-          onClick={() => setLoopPlayback((prevState) => !prevState)}
-          active={loopPlayback}
         />
       </ToolbarItem>
     </Toolbar>
