--- src/renderer/components/animator/PreviewToolbar/PreviewToolbar.css
@@ -1,5 +1,0 @@
-.preview-toolbar {
-  justify-content: center;
-  padding-bottom: 0;
-  height: var(--height-300);
-}
