+++ src/renderer/components/animator/PreviewToolbar/PreviewToolbar.tsx
@@ -1,12 +1,10 @@
+import { Flex, Group } from "@mantine/core";
 import { useContext } from "react";
 import CaptureContext from "../../../context/CaptureContext/CaptureContext";
 import PlaybackContext from "../../../context/PlaybackContext/PlaybackContext";
 import IconName from "../../common/Icon/IconName";
-import IconButton from "../../common/IconButton/IconButton";
-import Toolbar from "../../common/Toolbar/Toolbar";
-import ToolbarItem, { ToolbarItemAlign } from "../../common/ToolbarItem/ToolbarItem";
-import "./PreviewToolbar.css";
 import { UiActionIcon } from "../../ui/UiActionIcon/UiActionIcon";
+import { PreviewToolbarPlayback } from "./PreviewToolbarPlayback/PreviewToolbarPlayback";
 
 export const PreviewToolbar = (): JSX.Element => {
   const { captureImage } = useContext(CaptureContext);
@@ -20,12 +18,14 @@ export const PreviewToolbar = (): JSX.Element => {
   };
 
   return (
-    <Toolbar className="preview-toolbar">
-      <ToolbarItem align={ToolbarItemAlign.CENTER}>
-        <UiActionIcon icon={IconName.CAPTURE} onClick={handleClickCaptureButton} captureButton>
-          Capture Frame
-        </UiActionIcon>
-      </ToolbarItem>
-    </Toolbar>
+    <Group px="md">
+      <Flex flex={1}>
+        <PreviewToolbarPlayback />
+      </Flex>
+      <UiActionIcon icon={IconName.CAPTURE} onClick={handleClickCaptureButton} captureButton>
+        Capture Frame
+      </UiActionIcon>
+      <Flex flex={1}></Flex>
+    </Group>
   );
 };
